refactor: slim graph adapters and move shared policy to core
Move ToolApprovalHandler into cuga_agent_core, extract supervisor prepare/execute
nodes, and split the CugaLite adapter into focused modules. Address CodeRabbit
findings for step-limit ordering, find_tools fallback, StructuredTool.func
backfill, and Loguru formatting.
