+++ src/cuga/backend/cuga_graph/nodes/cuga_lite/adapter/__init__.py
@@ -1,0 +1,5 @@
+"""CugaLite graph adapter package."""
+
+from cuga.backend.cuga_graph.nodes.cuga_lite.adapter.graph_adapter import AgentGraphAdapter
+
+__all__ = ["AgentGraphAdapter"]
